# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "National Exam Card API",
        "description": "The National Exam Card API provides access to student's application for exam card data and process. ",
        "version": "0.1.0"
    },
    "paths": {
...
        "/marks": {
            "get": {
                "tags": [
                    "marks"
                ],
                "summary": "Get a list of all marks",
                "operationId": "c291194f941b3c5faadc35db74a64c23",
                "responses": {
                    "200": {
                        "description": "List of marks"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "post": {
                "tags": [
                    "marks"
                ],
                "summary": "Create a new marks",
                "operationId": "65e519ebab57e69351d438c0e4efb584",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Sample Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Sample content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Marks created successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    }
                }
            }
        },
        "/marks/{id}": {
            "get": {
                "tags": [
                    "marks"
                ],
                "summary": "Get a specific marks by ID",
                "operationId": "c46dda3250c2adc791e5071fa8c3b404",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the marks",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Marks details"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "put": {
                "tags": [
                    "marks"
                ],
                "summary": "Update an existing marks",
                "operationId": "17238cbdb5bae011ba7c8cb7dd373166",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the marks",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Updated Title"
                                    },
                                    "content": {
                                        "type": "string",
                                        "example": "Updated content"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 1
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Marks updated successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "marks"
                ],
                "summary": "Delete a marks by ID",
                "operationId": "4e96fe44070fb218bf151fcb99762ae9",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the marks",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Marks deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/notification": {
            "get": {
                "tags": [
                    "notification"
                ],
                "summary": "Get a list of all notifications",
                "operationId": "374cf5759ae311f38666cfd58baeeee3",
                "responses": {
                    "200": {
                        "description": "List of notifications"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
        "/notification/{id}": {
            "delete": {
                "tags": [
                    "notification"
                ],
                "summary": "Delete a notification by ID",
                "operationId": "bd79e4667720f473600515b005c5784a",
                "parameters": [
                    {
                        "name": "id",
                        "in": "path",
                        "description": "ID of the notification",
                        "required": true,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "204": {
                        "description": "Notification deleted successfully"
                    },
                    "401": {
                        "description": "Unauthorized"
                    },
                    "404": {
                        "description": "Not Found"
                    }
                }
            }
        },
...
    },
...
}